feature/routing: move all modules from folder 'pages' to folder 'modules'
