+++ modules/dashboard/client/dashboard.js
